Split workflow reconnection graph helper
