+++ packages/twenty-front/src/modules/workflow/workflow-diagram/utils/wouldReconnectWorkflowEdgeCreateCycle.ts
@@ -1,29 +1,7 @@
-import { type WorkflowAction } from '@/workflow/types/Workflow';
-import { getPreviousSteps } from '@/workflow/workflow-steps/utils/getWorkflowPreviousSteps';
-import { isDefined } from 'twenty-shared/utils';
-
-type WorkflowFlow = {
-  steps: WorkflowAction[] | null;
-};
-
-export const getWorkflowReconnectionForbiddenTargetStepIds = ({
-  flow,
-  sourceStepId,
-}: {
-  flow: WorkflowFlow;
-  sourceStepId: string;
-}): Set<string> => {
-  const steps = flow.steps ?? [];
-  const sourceStep = steps.find((step) => step.id === sourceStepId);
-
-  if (!isDefined(sourceStep)) {
-    return new Set<string>();
-  }
-
-  return new Set(
-    getPreviousSteps({ steps, currentStep: sourceStep }).map((step) => step.id),
-  );
-};
+import {
+  getWorkflowReconnectionForbiddenTargetStepIds,
+  type WorkflowFlow,
+} from '@/workflow/workflow-diagram/utils/getWorkflowReconnectionForbiddenTargetStepIds';
 
 export const wouldReconnectWorkflowEdgeCreateCycle = ({
   flow,
